+++ kbl24/src/main/java/com/foursoft/harness/kbl/v24/HasIdentification.java
@@ -23,12 +23,11 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.kblmodel.kbl24;
+package com.foursoft.harness.kbl.v24;
 
-import java.util.List;
+public interface HasIdentification {
 
-public interface HasAliasId {
+    String getId();
 
-    List<KblAliasIdentification> getAliasIds();
-
+    void setId(String id);
 }
