+++ kbl24/src/main/java/com/foursoft/harness/kbl/v24/HasPart.java
@@ -23,11 +23,10 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.kblmodel.kbl24;
+package com.foursoft.harness.kbl.v24;
 
-public interface HasIdentification {
+public interface HasPart {
 
-    String getId();
+    KblPart getPart();
 
-    void setId(String id);
 }
