+++ kbl24/src/main/java/com/foursoft/harness/kbl/v24/HasPlacement.java
@@ -23,7 +23,7 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.kblmodel.kbl24;
+package com.foursoft.harness.kbl.v24;
 
 public interface HasPlacement {
