+++ kbl24/src/main/java/com/foursoft/harness/kbl/v24/HasReferenceElement.java
@@ -23,11 +23,12 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.kblmodel.kbl24;
+package com.foursoft.harness.kbl.v24;
 
 import java.util.List;
 
-public interface HasInstallationInformation {
+public interface HasReferenceElement {
 
-    List<KblInstallationInstruction> getInstallationInformations();
+    List<ReferenceElement> getReferenceElement();
+
 }
